+++ solagent-plugins/pyth/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
 name = "solagent-plugin-pyth"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2021"
 authors = ["zTgx <beautifularea@gmail.com>"]
 repository = "https://github.com/zTgx/solagent.rs"
 keywords = ["solagent", "plugin", "pyth"]
 license = "Apache-2.0"
-description = "solagent.rs plugin pyth"
+description = "solagent plugin pyth"
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
